feat(suite-native): unify connect device handlers
# Conflicts:
#	suite-native/module-home/src/screens/HomeScreen/components/EmptyPortfolioCrossroads.tsx
